v2: Ignore other devices (#48)
An early design oversight has lead to a v2 release. The
`connectedDevice` event performs no checking to ensure the device that
connected is one the library cares about. I never tested the scenario of
connecting something other than a bunch of laser printers at the same
time until now. Since that required an API change I figured I'd add as
much API-breaking-changes in one go as I could.

Anyway.

This includes a switch to Vite, which I've found a VERY pleasant
ecosystem to work within for building the package.

## Breaking changes:

* USBDeviceManager is the new manager class.
* LabelPrinter is the new Printer class.
* A few other ancillary classes became interfaces and names changes
slightly. The types should be self-explanatory to follow.