Split things out for more testability (#59)
With the previous changes in place it's easier to make more of these
functions pure and break them out into separate modules. Writing more
tests against the smaller functions will be easier now.

This also moves the initial message handling function down to the base
CommandSet class, with a delegate being registered in the child language
classes. We also expose the current config to the message handling
delegate as there are some EPL messages that may benefit from the added
context.